+++ packages/python/openproblems/src/openproblems/project/read_viash_config.py
@@ -1,6 +1,6 @@
 import os
 from ..io.read_nested_yaml import read_nested_yaml
-from ..project.find_project_root import find_project_root
+from .find_project_root import find_project_root
 
 def read_viash_config(target_config_path, project_root_dir=None):
     """