+++ packages/python/openproblems/src/openproblems/viash/__init__.py
@@ -1,5 +1,0 @@
-from .read_viash_config import read_viash_config
-
-__all__ = [
-    "read_viash_config",
-]